chore: minimize mod replace
