+++ account_export_test.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package weshnet
 
 import (
