+++ contact_request_manager_test.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package weshnet
 
 import (
