+++ internal/handshake/handshake_util_test.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package handshake
 
 import (
