+++ pkg/ipfsutil/mobile/node.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package node
 
 import (
