+++ pkg/ipfsutil/repo.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package ipfsutil
 
 import (
