+++ pkg/secretstore/secret_store_messages_test.go
@@ -1,3 +1,5 @@
+//go:build !js
+
 package secretstore_test
 
 import (
